requirements.txt
scikit-learn
pandas
numpy
xgboost
lightgbm
fastapi
uvicorn
pydantic
pyyaml
pytest
joblib
httpx
